requirements.txt
torch>=2.1.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0
huggingface-hub>=0.23.0
bitsandbytes>=0.43.0
gradio>=4.20.0
Pillow>=10.0.0
pyyaml>=6.0
sentencepiece>=0.2.0
protobuf>=4.25.0
qwen-vl-utils>=0.0.8
